requirements/train.txt
torch==1.12.1
torchvision==0.13.1
diffusers==0.4.1
accelerate==0.17.1
transformers==4.24.0
ftfy
tensorboard
Jinja2
wandb
matplotlib
scikit-image
datasets
setuptools==59.5.0
mmpycocotools==12.0.3
opencv-python
bbox_visualizer
mmcv-full==1.7.0
mmdetection==2.25.3
modelcards